refactor: Split the lease order into separate files
